requirements.txt
python-pptx>=0.6.21
google-generativeai>=0.8.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
